backend/requirements.txt
flask==3.1.0
flask-cors==6.0.1
scikit-learn==1.6.1
seaborn==0.13.2
pandas==2.2.3
numpy==2.2.3
regex==2025.9.1
requests==2.32.5
beautifulsoup4==4.14.2
python-whois==0.9.6
dnspython==2.8.0
joblib==1.4.2
lightgbm==4.6.0
matplotlib==3.10.1
urllib3==2.4.0
whois==1.20240129.2
xgboost==3.0.0